add all features and finish the project

+++ app/Http/Controllers/front/banka/indexController.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers\front\banka;
 
 use App\Http\Controllers\Controller;
 use App\Models\Banka;
+use App\Models\Logger;
+use App\Models\UserPermission;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Redirect;
 use Yajra\DataTables\DataTables;
 
 class indexController extends Controller
@@ -13,6 +16,11 @@ class indexController extends Controller
     public function index()
     {
         return view('front.banka.index');
+
+    }
+    public function transfer()
+    {
+        return view('front.banka.transfer');
 
     }
 
@@ -24,14 +32,15 @@ class indexController extends Controller
     public function store(Request $request)
     {
         $all = $request->except('_token');
-
         $c1 = Banka::where('iban',$all['iban'])->count();
         $c2 = Banka::where('hesapNo',$all['hesapNo'])->count();
+
 
         if($c1 == 0 && $c2 == 0){
 
             $create = Banka::create($all);
             if ($create) {
+                Logger::Insert("Banka Eklendi","Banka");
                 return redirect()->back()->with('status', 'Banka Ekleme İşlemi Başarılı');
 
             } else {
@@ -53,12 +62,17 @@ class indexController extends Controller
 
     public function edit($id)
     {
-        $c = Banka::where('id', $id)->count();
-        if ($c != 0) {
-            $data = Banka::where('id', $id)->get();
-            return view('front.banka.edit', ['data' => $data]);
-        } else {
-            return redirect('/');
+        if(!UserPermission::getMyControl(6)) {
+            Redirect::to('/')->with('statusDanger','Talep reddedildi. Erişim iznine sahip değilsiniz.')->send();
+        }
+        else {
+            $c = Banka::where('id', $id)->count();
+            if ($c != 0) {
+                $data = Banka::where('id', $id)->get();
+                return view('front.banka.edit', ['data' => $data]);
+            } else {
+                return redirect('/');
+            }
         }
 
     }
@@ -74,6 +88,7 @@ class indexController extends Controller
 
             $update = Banka::where('id', $id)->update($all);
             if ($update) {
+                Logger::Insert($data[0]['ad']." Bankası Düzenlendi","Banka");
                 return redirect()->back()->with('status', 'Banka Başarıyla Güncellendi');
             } else {
                 return redirect()->back()->with('status', 'Banka Güncellenemedi');
@@ -85,14 +100,20 @@ class indexController extends Controller
 
     public function delete($id)
     {
-        $c = Banka::where('id', $id)->count();
-        if ($c != 0) {
-            $data = Banka::where('id', $id)->get();
-            Banka::where('id', $id)->delete();
+        if(!UserPermission::getMyControl(7)) {
+            Redirect::to('/')->with('statusDanger','Talep reddedildi. Erişim iznine sahip değilsiniz.')->send();
+        }
+        else {
+            $c = Banka::where('id', $id)->count();
+            if ($c != 0) {
+                $data = Banka::where('id', $id)->get();
+                Logger::Insert($data[0]['ad'] . " Bankası Silindi", "Banka");
+                Banka::where('id', $id)->delete();
 
-            return redirect()->back()->with('status', 'Banka Başarıyla Silindi');
-        } else {
-            return redirect('/')->with('status', 'Banka Silinemedi');
+                return redirect()->back()->with('status', 'Banka Başarıyla Silindi');
+            } else {
+                return redirect('/')->with('status', 'Banka Silinemedi');
+            }
         }
     }
 
@@ -101,14 +122,49 @@ class indexController extends Controller
         $table = Banka::query();
         $data = DataTables::of($table)
             ->addColumn('edit', function ($table) {
-                return '<a href="' . route('banka.edit', ['id' => $table->id]) . '"><i class="feather feather-edit list-icon"></i></a>';
+                return '<a href="' . route('banka.edit', ['id' => $table->id]) . '"><i title="Düzenle" style="color:darkgreen" class="feather feather-edit list-icon"></i></a>';
             })
             ->addColumn('delete', function ($table) {
-                return '<a href="' . route('banka.delete', ['id' => $table->id]) . '"><i class="feather feather-trash-2 list-icon"></i></a>';
+                return '<a class="deleteButton" href="' . route('banka.delete', ['id' => $table->id]) . '"><i title="Sil" style="color:darkred" class="feather feather-trash-2 list-icon"></i></a>';
             })
             ->rawColumns(['edit', 'delete'])
             ->make(true);
         return $data;
+    }
+
+    public function transferYap(Request $request) {
+        $all = $request->except('_token');
+
+        $transferTutar = $all['tutar'];
+
+        $gonderenData = Banka::where('id',$all['gonderenHesap'])->get();
+        $yatirilanData = Banka::where('id',$all['yatirilanHesap'])->get();
+
+        $gonderenTutar = $gonderenData[0]['bakiye'] - $transferTutar;
+        $yatirilanTutar = $yatirilanData[0]['bakiye'] + $transferTutar;
+
+        if($gonderenData[0]['id'] == $yatirilanData[0]['id']) {
+            return redirect()->back()->with('statusDanger',"Aynı hesaplar arasında işlem yapılamaz.");
+        }
+
+        else if($gonderenData[0]['bakiye'] < $transferTutar) {
+            return redirect()->back()->with('statusDanger',"Gönderen hesapta yeterli bakiye yok.");
+        }
+        else {
+            $gonderenUpdate = Banka::where('id', $all['gonderenHesap'])->update(['bakiye' => $gonderenTutar]);
+            $yatirilanUpdate = Banka::where('id', $all['yatirilanHesap'])->update(['bakiye' => $yatirilanTutar]);
+
+            if ($gonderenUpdate && $yatirilanUpdate) {
+                Logger::Insert($gonderenData[0]['ad'] . " => " . $yatirilanData[0]['ad'] . " " . $transferTutar . " TL transfer yapıldı.", "Banka");
+                return redirect()->back()->with('status', $gonderenData[0]['ad'] . " hesabından " . $yatirilanData[0]['ad'] . " hesabına " . $transferTutar . " TL transfer yapıldı.");
+            } else {
+                return redirect()->back()->with('status', 'Transfer Yapılamadı.');
+            }
+
+        }
+
+
+
     }
